nexus-rag/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sentence-transformers
torch
tf-keras
numpy
groq